Moved bitfield into its own directory
