+++ CMakeLists.txt
@@ -32,7 +32,7 @@ link_directories(
 # tests
 add_executable(tests
     src/tests.cpp
-    src/bitfield_tests.cpp
+    src/bitfield/bitfield_tests.cpp
     src/manager_tests.cpp
     src/pubsub/eventmanager_tests.cpp
 )
